requirements.txt
annotated-types==0.6.0
anyio==4.2.0
blinker==1.7.0
click==8.1.7
exceptiongroup==1.2.0
Flask==3.0.2
gunicorn==21.2.0
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pyarrow==15.0.0
pydantic==2.6.1
pydantic-settings==2.2.0
pydantic_core==2.16.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
threadpoolctl==3.3.0
typing_extensions==4.9.0
tzdata==2024.1
uWSGI==2.0.24
Werkzeug==3.0.1
